Education section dark mode fixes
