+++ lib/common/widgets/education/articles/home_article_column.dart
@@ -3,18 +3,20 @@ import 'package:solution_challenge/common/widgets/education/articles/home_articl
 import 'package:solution_challenge/utils/constants/colors.dart';
 import 'package:solution_challenge/utils/constants/image_strings.dart';
 import 'package:solution_challenge/utils/constants/sizes.dart';
+import 'package:solution_challenge/utils/helpers/helper_functions.dart';
 
 class PHomeArticleColumn extends StatelessWidget {
   const PHomeArticleColumn({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final dark = PHelperFunctions.isDarkMode(context);
     return Container(
       width: 300,
       margin: const EdgeInsets.only(right: 20),
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(TSizes.productImageRadius),
-        color: TColors.satin,
+        color: dark ? TColors.myblack : TColors.satin,
       ),
       child: const Column(
         children: [
