+++ lib/features/education/screens/resources/widgets/education_appbar.dart
@@ -1,26 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:iconsax/iconsax.dart';
 import 'package:solution_challenge/common/widgets/appbar/appbar.dart';
 import 'package:solution_challenge/utils/constants/colors.dart';
+import 'package:solution_challenge/utils/helpers/helper_functions.dart';
 
-class PEducationAppBar extends StatelessWidget {
-  const PEducationAppBar({
+class PEducationAppbar extends StatelessWidget {
+  const PEducationAppbar({
     super.key,
   });
 
   @override
   Widget build(BuildContext context) {
+
+    final dark = PHelperFunctions.isDarkMode(context);
     return PAppBar(
       title: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          Text("Hey Beautiful,",style: Theme.of(context).textTheme.labelMedium!.apply(color: TColors.accent)),
-          Text("Username",style: Theme.of(context).textTheme.headlineMedium!.apply(color: Colors.white)),
+          Text("Hey Beautiful,",
+              style: Theme.of(context)
+                  .textTheme
+                  .labelMedium!
+                  .apply(color: dark ? Colors.white : TColors.rani)),
+          Text("Username",
+              style: Theme.of(context)
+                  .textTheme
+                  .headlineMedium!
+                  .apply(color: dark ? Colors.white : TColors.rani)),
         ],
       ),
-      actions: [
-        IconButton(onPressed: (){}, icon: const Icon(Iconsax.category, color: Colors.white,))
-      ],
+      // actions: [
+      // ],
     );
   }
 }
